Split `util/status.js` in two.
The two facilities provided by this module were divergent in terms of
functionality and dependency; notably, the routines now in
`print_status_util.js` have no dependency on `NodeGit` and require no on-disk
repositories to use or test.
