+++ node/test/util/status_util.js
@@ -37,10 +37,9 @@ const Rebase              = require("../../lib/util/rebase");
 const RepoAST             = require("../../lib/util/repo_ast");
 const RepoASTTestUtil     = require("../../lib/util/repo_ast_test_util");
 const RepoStatus          = require("../../lib/util/repo_status");
-const Status              = require("../../lib/util/status");
+const StatusUtil          = require("../../lib/util/status_util");
 const SubmoduleUtil       = require("../../lib/util/submodule_util");
 const SubmoduleConfigUtil = require("../../lib/util/submodule_config_util");
-const UserError           = require("../../lib/util/user_error");
 
 // test utilities
 
@@ -144,10 +143,9 @@ remapRepoStatus = function (status, commitMap, urlMap) {
     });
 };
 
-describe("Status", function () {
+describe("StatusUti", function () {
     const FILESTATUS       = RepoStatus.FILESTATUS;
     const RELATION         = RepoStatus.Submodule.COMMIT_RELATION;
-    const StatusDescriptor = Status.StatusDescriptor;
     const Submodule        = RepoStatus.Submodule;
  
     describe("test.remapSubmodule", function () {
@@ -290,671 +288,6 @@ describe("Status", function () {
             it(caseName, function () {
                 const result = remapRepoStatus(c.input, c.commitMap, c.urlMap);
                 assert.deepEqual(result, c.expected);
-            });
-        });
-    });
-
-    describe("StatusDescriptor", function () {
-        describe("constructor", function () {
-            it("breathing", function () {
-                const descriptor = new StatusDescriptor(FILESTATUS.ADDED,
-                                                        "foo",
-                                                        "bar");
-                assert.equal(descriptor.status, FILESTATUS.ADDED);
-                assert.equal(descriptor.path, "foo");
-                assert.equal(descriptor.detail, "bar");
-            });
-        });
-        describe("print", function () {
-            const cases = {
-                "basic": {
-                    des: new StatusDescriptor(FILESTATUS.ADDED, "foo", "bar"),
-                    check: /foo.*bar/,
-                },
-                "added": {
-                    des: new StatusDescriptor(FILESTATUS.ADDED, "x", "y"),
-                    check: /^new file/,
-                },
-                "modified": {
-                    des: new StatusDescriptor(FILESTATUS.MODIFIED, "x", "y"),
-                    check: /^modified/,
-                },
-                "deleted": {
-                    des: new StatusDescriptor(FILESTATUS.REMOVED, "x", "y"),
-                    check: /^deleted/,
-                },
-                "conflicted": {
-                    des: new StatusDescriptor(FILESTATUS.CONFLICTED, "x", "y"),
-                    check: /^conflicted/,
-                },
-                "renamed": {
-                    des: new StatusDescriptor(FILESTATUS.RENAMED, "x", "y"),
-                    check: /^renamed/,
-                },
-                "type changed": {
-                    des: new StatusDescriptor(FILESTATUS.TYPECHANGED,
-                                              "x",
-                                              "y"),
-                    check: /^type changed/,
-                },
-                "with color": {
-                    des: new StatusDescriptor(FILESTATUS.TYPECHANGED,
-                                              "x",
-                                              "y"),
-                    color: text => `RED${text}RED`,
-                    check: /RED.*RED/,
-                },
-                "with cwd": {
-                    des: new StatusDescriptor(FILESTATUS.ADDED, "x", "y"),
-                    cwd: "q",
-                    check: /\.\.\/x/,
-                },
-            };
-            Object.keys(cases).forEach(caseName => {
-                const c = cases[caseName];
-                it(caseName, function () {
-                    let color = c.color;
-                    if (undefined === color) {
-                        color = x => x;
-                    }
-                    let cwd = c.cwd;
-                    if (undefined === cwd) {
-                        cwd = "";
-                    }
-                    const result = c.des.print(color, cwd);
-                    assert.match(result, c.check);
-                });
-            });
-        });
-    });
-
-    describe("sortDescriptorsByPath", function () {
-        const cases = {
-            "trivial": {
-                input: [],
-                expected: [],
-            },
-            "one": {
-                input: [new StatusDescriptor(FILESTATUS.ADDED, "x", "y")],
-                expected: [new StatusDescriptor(FILESTATUS.ADDED, "x", "y")],
-            },
-            "a few": {
-                input: [
-                    new StatusDescriptor(FILESTATUS.ADDED, "b", "b"),
-                    new StatusDescriptor(FILESTATUS.ADDED, "a", "a"),
-                    new StatusDescriptor(FILESTATUS.ADDED, "z", "z"),
-                    new StatusDescriptor(FILESTATUS.ADDED, "y", "y"),
-                    new StatusDescriptor(FILESTATUS.ADDED, "q", "q"),
-                    new StatusDescriptor(FILESTATUS.ADDED, "s", "s"),
-                ],
-                expected: [
-                    new StatusDescriptor(FILESTATUS.ADDED, "a", "a"),
-                    new StatusDescriptor(FILESTATUS.ADDED, "b", "b"),
-                    new StatusDescriptor(FILESTATUS.ADDED, "q", "q"),
-                    new StatusDescriptor(FILESTATUS.ADDED, "s", "s"),
-                    new StatusDescriptor(FILESTATUS.ADDED, "y", "y"),
-                    new StatusDescriptor(FILESTATUS.ADDED, "z", "z"),
-                ],
-            },
-        };
-        Object.keys(cases).forEach(caseName => {
-            const c = cases[caseName];
-            it(caseName, function () {
-                const result = Status.sortDescriptorsByPath(c.input);
-                assert.deepEqual(result, c.expected);
-            });
-        });
-    });
-
-    describe("printStatusDescriptors", function () {
-        // This function is implemented in terms of `StatusDescriptor.print`,
-        // so we just need to test a few things: basic function, sorting, and
-        // coloring.
-
-        const cases = {
-            "trivial": {
-                descriptors: [],
-                check: /^$/
-            },
-            "one": {
-                descriptors: [
-                    new StatusDescriptor(FILESTATUS.ADDED, "x", "y"),
-                ],
-                check: /new.*x.*y.*\n$/,
-            },
-            "color": {
-                descriptors: [
-                    new StatusDescriptor(FILESTATUS.ADDED, "x", "y"),
-                ],
-                color: x => `BLUE${x}BLUE`,
-                check: /BLUE/,
-            },
-            "order": {
-                descriptors: [
-                    new StatusDescriptor(FILESTATUS.ADDED, "Z", "y"),
-                    new StatusDescriptor(FILESTATUS.ADDED, "X", "y"),
-                ],
-                check: /X.*\n.*Z/,
-            },
-            "cwd": {
-                descriptors: [
-                    new StatusDescriptor(FILESTATUS.ADDED, "Z/q", "y"),
-                ],
-                cwd: "Z",
-                check: /\sq/,
-            },
-        };
-        Object.keys(cases).forEach(caseName => {
-            const c = cases[caseName];
-            it(caseName, function () {
-                const color = c.color || (x => x);
-                const cwd = c.cwd || "";
-                const result = Status.printStatusDescriptors(c.descriptors,
-                                                             color,
-                                                             cwd);
-                assert.match(result, c.check);
-            });
-        });
-    });
-
-    describe("printUntrackedFiles", function () {
-        const cases = {
-            "trivial": {
-                input: [],
-                check: /^$/,
-            },
-            "one": {
-                input: ["foo"],
-                check: /\tfoo\n/,
-            },
-            "two": {
-                input: ["foo", "bar"],
-                check: /\tbar\n\tfoo\n/,
-            },
-            "in color": {
-                input: ["foo"],
-                color: (x => `a${x}b`),
-                check: /\tafoob\n/,
-            },
-            "relative": {
-                input: ["foo"],
-                cwd: "bar",
-                check: /\t\.\.\/foo\n/,
-            },
-        };
-        Object.keys(cases).forEach(caseName => {
-            const c = cases[caseName];
-            it(caseName, function () {
-                const cwd = c.cwd || "";
-                const color = c.color || (x => x);
-                const result = Status.printUntrackedFiles(c.input, color, cwd);
-                assert.match(result, c.check);
-            });
-        });
-    });
-
-    describe("listSubmoduleDescriptors", function () {
-        const cases = {
-            "trivial": {
-                status: new RepoStatus(),
-            },
-            "deleted": {
-                status: new RepoStatus({
-                    submodules: {
-                        x: new RepoStatus.Submodule({
-                            commitUrl: "a",
-                            commitSha: "1",
-                            indexStatus: FILESTATUS.REMOVED,
-                        }),
-                    },
-                }),
-                staged: [
-                    new StatusDescriptor(FILESTATUS.REMOVED, "x", "submodule"),
-                ],
-            },
-            "new url": {
-                status: new RepoStatus({
-                    submodules: {
-                        x: new RepoStatus.Submodule({
-                            indexSha: "y",
-                            commitSha: "y",
-                            indexShaRelation: RELATION.SAME,
-                            commitUrl: "a",
-                            indexUrl: "b",
-                        }),
-                    },
-                }),
-                staged: [
-                    new StatusDescriptor(FILESTATUS.MODIFIED,
-                                         "x",
-                                         "submodule, new url"),
-                ],
-            },
-            "new commits in index and new url": {
-                status: new RepoStatus({
-                    submodules: {
-                        x: new RepoStatus.Submodule({
-                            indexSha: "y",
-                            commitSha: "z",
-                            indexShaRelation: RELATION.AHEAD,
-                            commitUrl: "a",
-                            indexUrl: "b",
-                        }),
-                    },
-                }),
-                staged: [
-                    new StatusDescriptor(FILESTATUS.MODIFIED,
-                                         "x",
-                                         "submodule, new commits, new url"),
-                ],
-            },
-            "new commits in index": {
-                status: new RepoStatus({
-                    submodules: {
-                        x: new RepoStatus.Submodule({
-                            indexSha: "y",
-                            commitSha: "z",
-                            indexShaRelation: RELATION.AHEAD,
-                            commitUrl: "a",
-                            indexUrl: "a",
-                        }),
-                    },
-                }),
-                staged: [
-                    new StatusDescriptor(FILESTATUS.MODIFIED,
-                                         "x",
-                                         "submodule, new commits"),
-                ],
-            },
-            "new commits in workdir": {
-                status: new RepoStatus({
-                    submodules: {
-                        x: new RepoStatus.Submodule({
-                            indexSha: "y",
-                            commitSha: "y",
-                            indexShaRelation: RELATION.SAME,
-                            commitUrl: "a",
-                            indexUrl: "a",
-                            workdirShaRelation: RELATION.AHEAD,
-                            repoStatus: new RepoStatus({
-                                head: "z",
-                            }),
-                        }),
-                    },
-                }),
-                staged: [
-                    new StatusDescriptor(FILESTATUS.MODIFIED,
-                                         "x",
-                                         "submodule, new commits"),
-                ],
-            },
-            "behind in index": {
-                status: new RepoStatus({
-                    submodules: {
-                        x: new RepoStatus.Submodule({
-                            indexSha: "y",
-                            commitSha: "z",
-                            indexShaRelation: RELATION.BEHIND,
-                            commitUrl: "a",
-                            indexUrl: "a",
-                        }),
-                    },
-                }),
-                staged: [
-                    new StatusDescriptor(FILESTATUS.MODIFIED,
-                                         "x",
-                                         "submodule, on old commit"),
-                ],
-            },
-            "unrelated in index": {
-                status: new RepoStatus({
-                    submodules: {
-                        x: new RepoStatus.Submodule({
-                            indexSha: "y",
-                            commitSha: "z",
-                            indexShaRelation: RELATION.UNRELATED,
-                            commitUrl: "a",
-                            indexUrl: "a",
-                        }),
-                    },
-                }),
-                staged: [
-                    new StatusDescriptor(FILESTATUS.MODIFIED,
-                                         "x",
-                                         "submodule, on unrelated commit"),
-                ],
-            },
-            "unknown in index": {
-                status: new RepoStatus({
-                    submodules: {
-                        x: new RepoStatus.Submodule({
-                            indexSha: "y",
-                            commitSha: "z",
-                            indexShaRelation: RELATION.UNKNOWN,
-                            commitUrl: "a",
-                            indexUrl: "a",
-                        }),
-                    },
-                }),
-                staged: [
-                    new StatusDescriptor(FILESTATUS.MODIFIED,
-                                         "x",
-                                         "submodule, on unknown commit"),
-                ],
-            },
-            "new sub, no commit": {
-                status: new RepoStatus({
-                    submodules: {
-                        x: new RepoStatus.Submodule({
-                            indexStatus: FILESTATUS.ADDED,
-                        }),
-                    },
-                }),
-                workdir: [
-                    new StatusDescriptor(
-                                  FILESTATUS.ADDED,
-                                  "x",
-                                  "submodule, create commit or stage changes"),
-                ],
-            },
-            "new sub, no commit, open": {
-                status: new RepoStatus({
-                    submodules: {
-                        x: new RepoStatus.Submodule({
-                            indexStatus: FILESTATUS.ADDED,
-                            repoStatus: new RepoStatus(),
-                        }),
-                    },
-                }),
-                workdir: [
-                    new StatusDescriptor(
-                                  FILESTATUS.ADDED,
-                                  "x",
-                                  "submodule, create commit or stage changes"),
-                ],
-            },
-            "new sub, no commit but staged": {
-                status: new RepoStatus({
-                    submodules: {
-                        x: new RepoStatus.Submodule({
-                            indexStatus: FILESTATUS.ADDED,
-                            repoStatus: new RepoStatus({
-                                staged: { foo: FILESTATUS.ADDED },
-                            }),
-                        }),
-                    },
-                }),
-                staged: [
-                    new StatusDescriptor(
-                                  FILESTATUS.ADDED,
-                                  "x",
-                                  "submodule, newly created"),
-                ],
-            },
-//            "new sub staged and good": {
-//                status
-//            },
-        };
-        Object.keys(cases).forEach(caseName => {
-            const c = cases[caseName];
-            it(caseName, function () {
-                const staged = c.staged || [];
-                const workdir = c.workdir || [];
-                const untracked = c.untracked || [];
-                const result = Status.listSubmoduleDescriptors(c.status);
-                assert.deepEqual(result, {
-                    staged: staged,
-                    workdir: workdir,
-                    untracked: untracked,
-                });
-            });
-        });
-    });
-
-    describe("accumulateStatus", function () {
-        const Desc = StatusDescriptor;
-        const cases = {
-            "trivial": {
-                input: new RepoStatus(),
-                expected: {
-                    staged: [],
-                    workdir: [],
-                    untracked: [],
-                },
-            },
-            "staged": {
-                input: new RepoStatus({
-                    staged: { x: FILESTATUS.REMOVED },
-                }),
-                expected: {
-                    staged: [ new Desc(FILESTATUS.REMOVED, "x", "") ],
-                    workdir: [],
-                    untracked: [],
-                },
-            },
-            "workdir": {
-                input: new RepoStatus({
-                    workdir: { x: FILESTATUS.MODIFIED},
-                }),
-                expected: {
-                    staged: [],
-                    workdir: [new Desc(FILESTATUS.MODIFIED, "x", "")],
-                    untracked: [],
-                },
-            },
-            "untracked": {
-                input: new RepoStatus({
-                    workdir: { x: FILESTATUS.ADDED},
-                }),
-                expected: {
-                    staged: [],
-                    workdir: [],
-                    untracked: ["x"],
-                },
-            },
-            "all": {
-                input: new RepoStatus({
-                    staged: { x: FILESTATUS.REMOVED },
-                    workdir: { 
-                        y: FILESTATUS.MODIFIED,
-                        z: FILESTATUS.ADDED,
-                    },
-                }),
-                expected: {
-                    staged: [ new Desc(FILESTATUS.REMOVED, "x", "") ],
-                    workdir: [new Desc(FILESTATUS.MODIFIED, "y", "")],
-                    untracked: ["z"],
-                },
-            },
-            "submodule-specific": {
-                input: new RepoStatus({
-                    submodules: {
-                        s: new Submodule({
-                            commitUrl: "foo",
-                            commitSha: "a",
-                            indexUrl: "bar",
-                        }),
-                    },
-                }),
-                expected: {
-                    staged: [
-                        new Desc(FILESTATUS.MODIFIED,
-                                 "s",
-                                 "submodule, new url"),
-                    ],
-                    workdir: [],
-                    untracked: [],
-                },
-            },
-            "submodule rollup": {
-                input: new RepoStatus({
-                    submodules: {
-                        s: new Submodule({
-                            commitUrl: "1",
-                            commitSha: "2",
-                            repoStatus: new RepoStatus({
-                                staged: { x: FILESTATUS.REMOVED },
-                                workdir: { 
-                                    y: FILESTATUS.MODIFIED,
-                                    z: FILESTATUS.ADDED,
-                                },
-                            }),
-                        }),
-                    },
-                }),
-                expected: {
-                    staged: [ new Desc(FILESTATUS.REMOVED, "s/x", "") ],
-                    workdir: [new Desc(FILESTATUS.MODIFIED, "s/y", "")],
-                    untracked: ["s/z"],
-                },
-            },
-        };
-        Object.keys(cases).forEach(caseName => {
-            const c = cases[caseName];
-            it(caseName, function () {
-                const result = Status.accumulateStatus(c.input);
-                const sort = Status.sortDescriptorsByPath;
-                const sortedResult = {
-                    staged: sort(result.staged),
-                    workdir: sort(result.workdir),
-                    untracked: result.untracked.sort(),
-                };
-                assert.deepEqual(sortedResult, c.expected);
-            });
-        });
-    });
-
-    describe("printRebase", function () {
-        const cases = {
-            "basic": {
-                input: new Rebase("master", "xxx", "ffffffffffffffff"),
-                check: /rebase in progress/
-            },
-            "branch": {
-                input: new Rebase("master", "xxx", "ffffffffffffffff"),
-                check: /master/
-            },
-            "sha": {
-                input: new Rebase("master", "xxx", "ffffffffffffffff"),
-                check: /ffff/,
-            },
-        };
-        Object.keys(cases).forEach(caseName => {
-            const c = cases[caseName];
-            it(caseName, function () {
-                const result = Status.printRebase(c.input);
-                assert.match(result, c.check);
-            });
-        });
-    });
-
-    describe("printCurrentBranch", function () {
-        const cases = {
-            "normal": {
-                input: new RepoStatus({
-                    currentBranchName: "master",
-                    headCommit: "aaaaaaaaaaaaaaaa",
-                }),
-                check: /master/,
-            },
-            "detached": {
-                input: new RepoStatus({
-                    headCommit: "aaaaaaaaaaaaaaaa",
-                }),
-                check: /aaaa/,
-            },
-        };
-        Object.keys(cases).forEach(caseName => {
-            const c = cases[caseName];
-            it(caseName, function () {
-                const result = Status.printCurrentBranch(c.input);
-                assert.match(result, c.check);
-            });
-        });
-    });
-
-    describe("printRepoStatus", function () {
-        // Most of the logic for this method is implemented in terms of
-        // other methods that are already tested; we just need to validate that
-        // the results are chained together.
-
-        // TODO: more testing here
-
-        // We'll use repo `x` for printing.
-
-        const cases = {
-            "trivial": {
-                input: new RepoStatus({
-                    currentBranchName: "master",
-                }),
-                regex: /On branch.*master.*\n.*nothing to commit.*/,
-            },
-            "rebase": {
-                input: new RepoStatus({
-                    currentBranchName: "master",
-                    rebase: new Rebase("x", "y", "z"),
-                }),
-                regex: /rebas/,
-            },
-            "detached": {
-                input: new RepoStatus({
-                    headCommit: "ffffaaaaffffaaaa",
-                }),
-                regex: /detached/,
-            },
-            "dirty meta": {
-                input: new RepoStatus({
-                    currentBranchName: "master",
-                    staged: {
-                        qrst: FILESTATUS.ADDED,
-                    },
-                }),
-                regex: /.*qrst/,
-            },
-            "dirty sub": {
-                input: new RepoStatus({
-                    currentBranchName: "master",
-                    submodules: {
-                        qrst: new Submodule({
-                            repoStatus: new RepoStatus({
-                                staged: {
-                                    "x/y/z": FILESTATUS.MODIFIED,
-                                },
-                            }),
-                        }),
-                    },
-                }),
-                regex: /qrst\/x\/y\/z/,
-            },
-            "cwd": {
-                input: new RepoStatus({
-                    currentBranchName: "master",
-                    staged: {
-                        qrst: FILESTATUS.ADDED,
-                    },
-                }),
-                cwd: "u/v",
-                regex: /\.\.\/\.\.\/qrst/,
-            },
-            "untracked": {
-                input: new RepoStatus({
-                    currentBranchName: "master",
-                    workdir: { foo: FILESTATUS.ADDED },
-                }),
-                regex: /foo/,
-            },
-        };
-        Object.keys(cases).forEach(caseName => {
-            const c = cases[caseName];
-            it(caseName, function () {
-                const cwd = c.cwd || "";
-                const result = Status.printRepoStatus(c.input, cwd);
-                if (c.inverse) {
-                    assert.notMatch(result, c.regex);
-                }
-                else {
-                    assert.match(result, c.regex);
-                }
             });
         });
     });
@@ -1121,9 +454,9 @@ x=S:C2 a/b=c,a/c=d,t=u;H=2;I a/b,a/q=r,f=x;W a/b=q,a/c=f,a/y=g,f`,
                 const repo = w.repos.x;
                 const paths = c.paths || [];
                 const allUntracked = c.allUntracked || false;
-                const result = yield Status.getChanges(repo,
-                                                       paths,
-                                                       allUntracked);
+                const result = yield StatusUtil.getChanges(repo,
+                                                           paths,
+                                                           allUntracked);
                 const expected = {
                     staged: c.staged || {},
                     workdir: c.workdir || {},
@@ -1297,14 +630,15 @@ x=S:C2 a/b=c,a/c=d,t=u;H=2;I a/b,a/q=r,f=x;W a/b=q,a/c=f,a/y=g,f`,
                 const commitUrl = commitUrls.s || null;
                 const commitTree = yield commit.getTree();
                 const isVisible = yield SubmoduleUtil.isVisible(repo, "s");
-                const result = yield Status.getSubmoduleStatus("s",
-                                                               repo,
-                                                               indexUrl,
-                                                               commitUrl,
-                                                               index,
-                                                               commitTree,
-                                                               isVisible,
-                                                               getRepoStatus);
+                const result = yield StatusUtil.getSubmoduleStatus(
+                                                                "s",
+                                                                repo,
+                                                                indexUrl,
+                                                                commitUrl,
+                                                                index,
+                                                                commitTree,
+                                                                isVisible,
+                                                                getRepoStatus);
                 assert.instanceOf(result, RepoStatus.Submodule);
                 const mappedResult = remapSubmodule(result,
                                                     w.commitMap,
@@ -1564,145 +898,13 @@ x=S:C2 a/b=c,a/c=d,t=u;H=2;I a/b,a/q=r,f=x;W a/b=q,a/c=f,a/y=g,f`,
             const c = cases[caseName];
             it(caseName, co.wrap(function *() {
                 const w = yield RepoASTTestUtil.createMultiRepos(c.state);
-                const result = yield Status.getRepoStatus(w.repos.x,
-                                                          c.options);
+                const result = yield StatusUtil.getRepoStatus(w.repos.x,
+                                                              c.options);
                 assert.instanceOf(result, RepoStatus);
                 const mappedResult = remapRepoStatus(result,
                                                      w.commitMap,
                                                      w.urlMap);
                 assert.deepEqual(mappedResult, c.expected);
-            }));
-        });
-    });
-
-    describe("ensureClean", function () {
-        // We don't need to test the `isClean` functionality; it's already
-        // tested, just that it's called propertly.
-        // TODO: regex on error message.
-
-        // We will check the repo named `x`.
-        const cases = {
-            "trivial": {
-                state: "x=S",
-                fails: false,
-            },
-            "and a clean submodule": {
-                state: "a=S|x=S:C2-1 s=Sa:1;Bmaster=2;Os",
-                fails: false,
-            },
-            "dirty meta": {
-                state: "x=S:I foo=bar",
-                fails: true,
-            },
-            "and a dirty sub": {
-                state: "a=S|x=S:C2-1 s=Sa:1;Bmaster=2;Os W README.md=aaa",
-                fails: true,
-            },
-        };
-        Object.keys(cases).forEach(caseName => {
-            const c = cases[caseName];
-            it(caseName, co.wrap(function *() {
-                const w = yield RepoASTTestUtil.createMultiRepos(c.state);
-                const x = w.repos.x;
-                const xStat = yield Status.getRepoStatus(x);
-                try {
-                    Status.ensureClean(xStat);
-                    assert.equal(c.fails, false);
-                    return;                                           // RETURN
-                }
-                catch (e) {
-                    if (!(e instanceof UserError)) {
-                        throw e;
-                    }
-                }
-                assert(c.fails);
-            }));
-        });
-    });
-
-    describe("ensureConsistent", function () {
-        // TODO: check formatting of error message.
-
-        // We'll be checking the consistency of the repo named `x`.
-        const cases = {
-            "trivial": {
-                state: "x=S",
-                fails: false,
-            },
-            "good sub": {
-                state: "a=S|x=S:C2-1 s=Sa:1;Bmaster=2",
-                fails: false,
-            },
-            "sub with staged change": {
-                state: "a=S|x=S:I s=Sa:1",
-                fails: true,
-            },
-            "good open sub": {
-                state: "a=S|x=S:C2-1 s=Sa:1;Bmaster=2;Os Bmaster=1!*=master",
-                fails: false,
-            },
-            "sub with new commit in open repo": {
-                state: "a=S|x=S:C2-1 s=Sa:1;Bmaster=2;Os Bmaster=2!*=master",
-                fails: true,
-            },
-        };
-        Object.keys(cases).forEach(caseName => {
-            const c = cases[caseName];
-            it(caseName, co.wrap(function *() {
-                const w = yield RepoASTTestUtil.createMultiRepos(c.state);
-                const x = w.repos.x;
-                const xStat = yield Status.getRepoStatus(x);
-                let error = "";
-                try {
-                    Status.ensureConsistent(xStat);
-                    assert.equal(c.fails, false);
-                    return;                                           // RETURN
-                }
-                catch (e) {
-                    if (!(e instanceof UserError)) {
-                        throw e;
-                    }
-                    error = e.stack;
-                }
-                assert(c.fails, error);
-            }));
-        });
-    });
-
-    describe("ensureCleanAndConsistent", function () {
-        // We don't have to test much here; this method defers to `ensureClean`
-        // and `ensureConsistent`.
-
-        // Ensure the repo `x`.
-        const cases = {
-            "trivial": {
-                state: "x=S",
-                fails: false,
-            },
-            "unclean": {
-                state: "x=S:I foo=bar",
-                fails: true,
-            },
-        };
-        Object.keys(cases).forEach(caseName => {
-            const c = cases[caseName];
-            it(caseName, co.wrap(function *() {
-                const w = yield RepoASTTestUtil.createMultiRepos(c.state);
-                const x = w.repos.x;
-                let error = "";
-                const status = yield Status.getRepoStatus(x);
-                try {
-                    Status.ensureCleanAndConsistent(status);
-                    assert.equal(c.fails, false);
-                    return;                                           // RETURN
-                }
-                catch (e) {
-                    if (!(e instanceof UserError)) {
-                        throw e;
-                    }
-                    error = e.stack;
-                }
-                assert(c.fails, error);
             }));
         });
     });
